#38 Изменила <> на Integer

+++ src/main/java/ru/skypro/homework/repository/AdRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdRepository extends JpaRepository<AdEntity, Long> {
+public interface AdRepository extends JpaRepository<AdEntity, Integer> {
     List<AdEntity> findByAuthor(UserEntity author);
 
     // Остальные методы репозитория стандартные
